requirements.txt
gradio==5.49.1
fastapi
uvicorn
python-dotenv
pymongo
dnspython
cloudinary
requests
pillow
secure-smtplib